Changed line about NodeBB being free.
Changed line "The next generation forum software that's free and easy to use." to "NodeBB is next generation forum software. It's powerful, mobile-ready and easy to use."

The theory is that some people read this and assume there is a free version of SaaS hosting, when what we are talking about is the open source, self-hosted version.